カウンター画面 自動カウント処理をコンポーザブルに抽出
